# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MindVault AI Personal Workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List items with optional filters
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [note, file, reminder]
        - name: folderId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, completed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createItem
      tags: [items]
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get an item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateItem
      tags: [items]
      summary: Update an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/upload:
    post:
      operationId: uploadFile
      tags: [items]
      summary: Upload a file and create a file item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadFileBody"
      responses:
        "201":
          description: Created file item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # AI chat (Gemini-powered)
...
  /search:
    get:
      operationId: searchItems
      tags: [search]
      summary: Search items by query
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [note, file, reminder]
        - name: folderId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Stats
  /stats/workspace:
    get:
      operationId: getWorkspaceStats
      tags: [stats]
      summary: Get workspace overview stats
      responses:
        "200":
          description: Workspace stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stats/recent:
    get:
      operationId: getRecentItems
      tags: [stats]
      summary: Get recently created/modified items
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stats/reminders/upcoming:
    get:
      operationId: getUpcomingReminders
      tags: [stats]
      summary: Get upcoming reminders sorted by date
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Upcoming reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth schemas
...
    Item:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        folderId:
          type: ["integer", "null"]
        folderName:
          type: ["string", "null"]
        type:
          type: string
          enum: [note, file, reminder]
        title:
          type: string
        content:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        originalFilename:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        fileData:
          type: ["string", "null"]
        reminderAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [active, archived, completed]
        aiCategory:
          type: ["string", "null"]
        aiTags:
          type: array
          items:
            type: string
        aiConfidence:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - status
        - aiTags
        - createdAt
        - updatedAt
    CreateItemBody:
      type: object
      properties:
        type:
          type: string
          enum: [note, file, reminder]
        title:
          type: string
        content:
          type: ["string", "null"]
        folderId:
          type: ["integer", "null"]
        reminderAt:
          type: ["string", "null"]
      required:
        - type
        - title
    UpdateItemBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: ["string", "null"]
        folderId:
          type: ["integer", "null"]
        reminderAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived, completed]
    UploadFileBody:
      type: object
      properties:
        filename:
          type: string
        mimeType:
          type: string
        fileData:
          type: string
          description: Base64-encoded file data
        fileSize:
          type: integer
        folderId:
          type: ["integer", "null"]
      required:
        - filename
        - mimeType
        - fileData
        - fileSize

    # AI classification schemas
...
    WorkspaceStats:
      type: object
      properties:
        totalItems:
          type: integer
        totalNotes:
          type: integer
        totalFiles:
          type: integer
        totalReminders:
          type: integer
        totalFolders:
          type: integer
        pendingReminders:
          type: integer
        totalConversations:
          type: integer
      required:
        - totalItems
        - totalNotes
        - totalFiles
        - totalReminders
        - totalFolders
        - pendingReminders
        - totalConversations

    # Gemini chat schemas
...