# lib/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindVault AI Personal Workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateItemBodyType } from "./createItemBodyType";

export interface CreateItemBody {
  type: CreateItemBodyType;
  title: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  reminderAt?: string | null;
}