# lib/api-zod/src/generated/types/updateItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindVault AI Personal Workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateItemBodyStatus } from "./updateItemBodyStatus";

export interface UpdateItemBody {
  title?: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  reminderAt?: string | null;
  status?: UpdateItemBodyStatus;
}